#16 refactor: letterMessage 패키지명 수정
