+++ src/main/java/com/example/sharemind/letter/domain/Letter.java
@@ -5,9 +5,9 @@ import com.example.sharemind.customer.domain.Customer;
 import com.example.sharemind.global.common.BaseEntity;
 import com.example.sharemind.global.content.ConsultCategory;
 import com.example.sharemind.letter.content.LetterStatus;
-import com.example.sharemind.nonRealtimeMessage.content.LetterMessageType;
-import com.example.sharemind.nonRealtimeMessage.exception.LetterMessageErrorCode;
-import com.example.sharemind.nonRealtimeMessage.exception.LetterMessageException;
+import com.example.sharemind.letterMessage.content.LetterMessageType;
+import com.example.sharemind.letterMessage.exception.LetterMessageErrorCode;
+import com.example.sharemind.letterMessage.exception.LetterMessageException;
 import jakarta.persistence.*;
 import lombok.*;
