+++ src/main/java/com/example/sharemind/letterMessage/application/LetterMessageService.java
@@ -1,8 +1,8 @@
-package com.example.sharemind.nonRealtimeMessage.application;
+package com.example.sharemind.letterMessage.application;
 
 import com.example.sharemind.customer.domain.Customer;
-import com.example.sharemind.nonRealtimeMessage.dto.request.LetterMessageCreateRequest;
-import com.example.sharemind.nonRealtimeMessage.dto.request.LetterMessageUpdateRequest;
+import com.example.sharemind.letterMessage.dto.request.LetterMessageCreateRequest;
+import com.example.sharemind.letterMessage.dto.request.LetterMessageUpdateRequest;
 
 public interface LetterMessageService {
     void createLetterMessage(LetterMessageCreateRequest letterMessageCreateRequest, Customer customer);
