+++ src/main/java/com/example/sharemind/letterMessage/domain/LetterMessage.java
@@ -1,11 +1,11 @@
-package com.example.sharemind.nonRealtimeMessage.domain;
+package com.example.sharemind.letterMessage.domain;
 
 import com.example.sharemind.global.common.BaseEntity;
 import com.example.sharemind.letter.content.LetterStatus;
 import com.example.sharemind.letter.domain.Letter;
-import com.example.sharemind.nonRealtimeMessage.content.LetterMessageType;
-import com.example.sharemind.nonRealtimeMessage.exception.LetterMessageErrorCode;
-import com.example.sharemind.nonRealtimeMessage.exception.LetterMessageException;
+import com.example.sharemind.letterMessage.content.LetterMessageType;
+import com.example.sharemind.letterMessage.exception.LetterMessageErrorCode;
+import com.example.sharemind.letterMessage.exception.LetterMessageException;
 import jakarta.persistence.*;
 import lombok.*;
