+++ src/main/java/com/example/sharemind/letterMessage/exception/LetterMessageErrorCode.java
@@ -1,4 +1,4 @@
-package com.example.sharemind.nonRealtimeMessage.exception;
+package com.example.sharemind.letterMessage.exception;
 
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
